local workers without reboot now correctly see the update of ComfyUI paths
