+++ visionatrix/alembic/versions/52d98e8fd54f_added_crc32_to_global_settings.py
@@ -1,0 +1,30 @@
+"""Added CRC32 to global settings
+
+Revision ID: 52d98e8fd54f
+Revises: 3bd1c6a2efd7
+Create Date: 2024-12-03 12:14:36.346048
+
+"""
+
+from collections.abc import Sequence
+
+import sqlalchemy as sa
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "52d98e8fd54f"
+down_revision: str | None = "3bd1c6a2efd7"
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column("global_settings", sa.Column("crc32", sa.BigInteger(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("global_settings", "crc32")
+    # ### end Alembic commands ###
