+++ visionatrix/routes/settings_comfyui.py
@@ -1,22 +1,14 @@
 import json
-import logging
-from datetime import datetime
 from json import loads
-from pathlib import Path
+from zlib import crc32
 
 from fastapi import APIRouter, Body, HTTPException, Request, status
 
-from .. import options
-from ..comfyui_wrapper import add_model_folder_path, get_folder_names_and_paths
+from .. import comfyui_wrapper, settings_comfyui
 from ..db_queries import get_global_setting, set_global_setting
-from ..pydantic_models import (
-    ComfyUIFolderPath,
-    ComfyUIFolderPathDefinition,
-    ComfyUIFolderPaths,
-)
+from ..pydantic_models import ComfyUIFolderPathDefinition, ComfyUIFolderPaths
 from .helpers import require_admin
 
-LOGGER = logging.getLogger("visionatrix")
 ROUTER = APIRouter(prefix="/settings/comfyui", tags=["settings"])
 
 
@@ -36,7 +28,7 @@ def comfyui_get_folders_paths(request: Request) -> ComfyUIFolderPaths:
     if comfyui_folders_setting := get_global_setting("comfyui_folders", True):
         comfyui_folders = [ComfyUIFolderPathDefinition.model_validate(i) for i in loads(comfyui_folders_setting)]
 
-    return compute_folder_paths(comfyui_folders)
+    return settings_comfyui.compute_folder_paths(comfyui_folders)
 
 
 @ROUTER.post(
@@ -54,31 +46,16 @@ def add_folder_path(
     """
     require_admin(request)
 
-    comfyui_folders: list[ComfyUIFolderPathDefinition] = []
-    if comfyui_folders_setting := get_global_setting("comfyui_folders", True):
-        comfyui_folders = [ComfyUIFolderPathDefinition.model_validate(i) for i in loads(comfyui_folders_setting)]
+    try:
+        comfyui_folders = settings_comfyui.add_folder_path(body)
+    except ValueError as e:
+        raise HTTPException(status.HTTP_400_BAD_REQUEST, str(e)) from None
 
-    absolute_new_path = body.path
-    if not Path(absolute_new_path).is_absolute():
-        absolute_new_path = Path(options.COMFYUI_DIR).joinpath(body.path).resolve()
-
-    for custom_folder in comfyui_folders:
-        custom_folder_path = Path(custom_folder.path)
-        if not custom_folder_path.is_absolute():
-            custom_folder_path = Path(options.COMFYUI_DIR).joinpath(custom_folder.path).resolve()
-        if body.folder_key == custom_folder.folder_key and absolute_new_path == custom_folder_path:
-            raise HTTPException(
-                status.HTTP_400_BAD_REQUEST,
-                f"The folder path '{body.path}' already exists under the key '{body.folder_key}'.",
-            )
-
-    add_model_folder_path(body.folder_key, str(absolute_new_path), body.is_default)
-
-    comfyui_folders.append(body)
-    set_global_setting(
-        "comfyui_folders", json.dumps([i.model_dump(mode="json") for i in comfyui_folders]), sensitive=True
-    )
-    return compute_folder_paths(comfyui_folders)
+    raw_data = json.dumps([i.model_dump(mode="json") for i in comfyui_folders])
+    raw_data_crc32 = crc32(raw_data.encode("utf-8"))
+    comfyui_wrapper.COMFYUI_FOLDERS_SETTING_CRC32 = raw_data_crc32
+    set_global_setting("comfyui_folders", raw_data, sensitive=True, crc32=raw_data_crc32)
+    return settings_comfyui.compute_folder_paths(comfyui_folders)
 
 
 @ROUTER.delete(
@@ -97,44 +74,19 @@ def remove_folder_path(
     """
     require_admin(request)
 
-    comfyui_folders: list[ComfyUIFolderPathDefinition] = []
-    if comfyui_folders_setting := get_global_setting("comfyui_folders", True):
-        comfyui_folders = [ComfyUIFolderPathDefinition.model_validate(i) for i in loads(comfyui_folders_setting)]
-
-    absolute_path = Path(path)
-    if not absolute_path.is_absolute():
-        absolute_path = Path(options.COMFYUI_DIR).joinpath(path).resolve()
-
-    updated_folders = []
-    folder_found = False
-    for custom_folder in comfyui_folders:
-        custom_folder_path = Path(custom_folder.path)
-        if not custom_folder_path.is_absolute():
-            custom_folder_path = Path(options.COMFYUI_DIR).joinpath(custom_folder.path).resolve()
-        if custom_folder.folder_key == folder_key and custom_folder_path == absolute_path:
-            folder_found = True
-        else:
-            updated_folders.append(custom_folder)
-
-    if not folder_found:
+    try:
+        comfyui_folders = settings_comfyui.remove_folder_path(folder_key, path)
+    except ValueError as e:
         raise HTTPException(
-            status.HTTP_404_NOT_FOUND,
-            f"The folder path '{path}' under the key '{folder_key}' was not found.",
-        )
+            status.HTTP_400_BAD_REQUEST,
+            str(e),
+        ) from None
 
-    folder_names_and_paths = get_folder_names_and_paths()
-    if folder_key in folder_names_and_paths:
-        folder_paths, _ = folder_names_and_paths[folder_key]
-        try:
-            folder_paths.remove(str(absolute_path))
-            LOGGER.info("Removed path '%s' from ComfyUI key '%s'.", absolute_path, folder_key)
-        except ValueError:
-            LOGGER.error("Path '%s' not found in ComfyUI key '%s'.", absolute_path, folder_key)
-
-    set_global_setting(
-        "comfyui_folders", json.dumps([i.model_dump(mode="json") for i in updated_folders]), sensitive=True
-    )
-    return compute_folder_paths(updated_folders)
+    raw_data = json.dumps([i.model_dump(mode="json") for i in comfyui_folders])
+    raw_data_crc32 = crc32(raw_data.encode("utf-8"))
+    comfyui_wrapper.COMFYUI_FOLDERS_SETTING_CRC32 = raw_data_crc32
+    set_global_setting("comfyui_folders", raw_data, sensitive=True, crc32=raw_data_crc32)
+    return settings_comfyui.compute_folder_paths(comfyui_folders)
 
 
 @ROUTER.post(
@@ -158,90 +110,11 @@ def autoconfigure_model_folders(
             "ComfyUI folder settings are not empty.",
         )
 
-    paths_to_preconfigure = {
-        "checkpoints": "checkpoints",
-        "text_encoders": "text_encoders",
-        "clip_vision": "clip_vision",
-        "controlnet": "controlnet",
-        "diffusion_models": "diffusion_models",
-        "diffusers": "diffusers",
-        "ipadapter": "ipadapter",
-        "instantid": "instantid",
-        "loras": ["loras", "photomaker"],
-        "photomaker": "photomaker",
-        "sams": "sams",
-        "ultralytics": "ultralytics",
-        "unet": "unet",
-        "upscale_models": "upscale_models",
-        "vae": "vae",
-        "vae_approx": "vae_approx",
-        "pulid": "pulid",
-    }
-
-    comfyui_folders = []
-    for folder_key, subpaths in paths_to_preconfigure.items():
-        if isinstance(subpaths, str):
-            subpaths = [subpaths]
-        for subpath in subpaths:
-            path = str(Path(models_dir).joinpath(subpath))
-            folder_def = ComfyUIFolderPathDefinition(
-                folder_key=folder_key,
-                path=path,
-                is_default=True,
-            )
-
-            absolute_new_path = folder_def.path
-            if not Path(absolute_new_path).is_absolute():
-                absolute_new_path = Path(options.COMFYUI_DIR).joinpath(folder_def.path).resolve()
-
-            add_model_folder_path(folder_def.folder_key, str(absolute_new_path), folder_def.is_default)
-            comfyui_folders.append(folder_def)
+    comfyui_folders = settings_comfyui.autoconfigure_model_folders(models_dir)
 
     # Update the global setting once with all folder paths
-    set_global_setting(
-        "comfyui_folders",
-        json.dumps([i.model_dump(mode="json") for i in comfyui_folders]),
-        sensitive=True,
-    )
-
-    return compute_folder_paths(comfyui_folders)
-
-
-def compute_folder_paths(comfyui_folders: list[ComfyUIFolderPathDefinition]) -> ComfyUIFolderPaths:
-    folder_data = {}
-    for key, (paths, _) in get_folder_names_and_paths().items():
-        folder_data[key] = []
-        for folder in paths:
-            folder_path = Path(folder)
-            total_size = 0
-            create_time = datetime.fromtimestamp(0.0)
-            if folder_path.exists() and folder_path.is_dir():
-                try:
-                    create_time = datetime.fromtimestamp(folder_path.stat().st_birthtime)
-                except AttributeError:
-                    # Fall back if `st_birthtime` is not available (e.g., on some Linux filesystems)
-                    create_time = datetime.fromtimestamp(folder_path.stat().st_ctime)
-
-                total_size = sum(f.stat().st_size for f in folder_path.rglob("*") if f.is_file())
-
-            readonly = True
-            is_default = False
-            for custom_folder in comfyui_folders:
-                custom_folder_path = Path(custom_folder.path)
-                if not custom_folder_path.is_absolute():
-                    custom_folder_path = Path(options.COMFYUI_DIR).joinpath(custom_folder.path).resolve()
-                if key == custom_folder.folder_key and folder_path == custom_folder_path:
-                    readonly = False
-                    is_default = custom_folder.is_default
-                    break
-
-            folder_data[key].append(
-                ComfyUIFolderPath(
-                    readonly=readonly,
-                    full_path=str(folder_path),
-                    is_default=is_default,
-                    create_time=create_time,
-                    total_size=total_size,
-                )
-            )
-    return ComfyUIFolderPaths(folders=folder_data)
+    raw_data = json.dumps([i.model_dump(mode="json") for i in comfyui_folders])
+    raw_data_crc32 = crc32(raw_data.encode("utf-8"))
+    comfyui_wrapper.COMFYUI_FOLDERS_SETTING_CRC32 = raw_data_crc32
+    set_global_setting("comfyui_folders", raw_data, sensitive=True, crc32=raw_data_crc32)
+    return settings_comfyui.compute_folder_paths(comfyui_folders)
